backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-multipart==0.0.9
requests==2.32.3
pydantic==2.7.4
alembic==1.13.1
ffmpeg-python==0.2.0
pydantic-settings
python-jose
python-dotenv
werkzeug
numpy
pandas
scikit-learn
sentence-transformers
